{{cookiecutter.project_slug}}/requirements.txt
uvicorn==0.21.1
fastapi==0.95.1
requests==2.28.2
xmltodict==0.13.0
sqlalchemy==2.0.9
celery==5.2.7
alembic==1.10.3
starlette==0.26.1
pydantic==1.10.7
python-dotenv==1.0.0
